+++ source/block/Lava.cpp
@@ -7,8 +7,9 @@
 
 static blockTexture *tex_lava;
 
-static void render(int xPos, int yPos, int zPos, unsigned char pass) {
-	if (pass == 0) return;
+static void render(s16 xPos, s16 yPos, s16 zPos, unsigned char pass) {
+	if (pass == 0)
+		return;
 	Render::drawLiquidBlock(xPos, yPos, zPos, tex_lava);
 }
