+++ source/block/Water.cpp
@@ -11,8 +11,9 @@ static blockTexture *tex_water;
 
 inline void drawWaterBlock(int xPos, int yPos, int zPos);
 
-static void render(int xPos, int yPos, int zPos, unsigned char pass) {
-	if (pass == 0) return;
+static void render(s16 xPos, s16 yPos, s16 zPos, unsigned char pass) {
+	if (pass == 0)
+		return;
 	Render::drawLiquidBlock(xPos, yPos, zPos, tex_water);
 }
